Update README.md
Mind Dump is the name we chose because it reflects the safe space we all need when emotions pile up. Like a personal outlet, it represents our shared goal to support each other through the ups and downs of student life. The name and logo symbolize the importance of letting things out, one thought at a time—without judgment.